:bug: 暂时修复插入后返回主键为Long类型造成的转型异常

+++ src/main/java/com/tale/service/MetasService.java
@@ -104,7 +104,8 @@ public class MetasService {
             metas.setSlug(name);
             metas.setName(name);
             metas.setType(type);
-            mid = metas.save();
+            //TODO 下策，无法解决save后返回long导致转型异常的BUG，暂时如此解决
+            mid =  ((Long)metas.save()).intValue();
         }
         if (mid != 0) {
             long count = new Relationships().where("cid", cid).and("mid", mid).count();
